# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for managing users, tasks, and task statistics.",
        "title": "Notes app API",
        "contact": {},
        "version": "1.0"
    },
    "host": "127.0.0.1:5050",
    "basePath": "/api/v1",
    "paths": {
...
        "/tasks": {
            "get": {
                "description": "Returns tasks with optional filtering by `user_id` (UUID) and optional pagination.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "List tasks",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Filter tasks by author user ID (UUID)",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tasks list",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_features_tasks_transport_http.taskDTOResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_rod1kutzyy_task-manager-app_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_rod1kutzyy_task-manager-app_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new task in the system.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Create a task",
                "parameters": [
                    {
                        "description": "Task creation payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_features_tasks_transport_http.createTaskRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created task",
                        "schema": {
                            "$ref": "#/definitions/internal_features_tasks_transport_http.createTaskResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_rod1kutzyy_task-manager-app_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Author not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_rod1kutzyy_task-manager-app_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_rod1kutzyy_task-manager-app_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/tasks/{id}": {
            "get": {
                "description": "Returns a single task by UUID identifier.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Get a task by ID",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Task ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task found",
                        "schema": {
                            "$ref": "#/definitions/internal_features_tasks_transport_http.getTaskResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_rod1kutzyy_task-manager-app_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Task not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_rod1kutzyy_task-manager-app_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_rod1kutzyy_task-manager-app_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a task by UUID identifier.",
                "tags": [
                    "tasks"
                ],
                "summary": "Delete a task",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Task ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Task deleted"
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_rod1kutzyy_task-manager-app_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Task not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_rod1kutzyy_task-manager-app_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_rod1kutzyy_task-manager-app_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Updates task fields using three-state semantics for each field.\n1. Field is omitted: the value is not changed.\n2. Field has a value: the value is updated.\n3. Field is explicitly null: the value is cleared (set to NULL).\nConstraints: `title` and `completed` cannot be set to null.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Partially update a task",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Task ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Task patch payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_features_tasks_transport_http.patchTaskRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated task",
                        "schema": {
                            "$ref": "#/definitions/internal_features_tasks_transport_http.patchTaskResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_rod1kutzyy_task-manager-app_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Task not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_rod1kutzyy_task-manager-app_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_rod1kutzyy_task-manager-app_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_rod1kutzyy_task-manager-app_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/users/{id}": {
            "get": {
                "description": "Returns a single user by UUID identifier.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get a user by ID",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "User ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User found",
                        "schema": {
                            "$ref": "#/definitions/internal_features_users_transport_http.getUserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_rod1kutzyy_task-manager-app_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_rod1kutzyy_task-manager-app_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_rod1kutzyy_task-manager-app_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a user by UUID identifier.",
                "tags": [
                    "users"
                ],
                "summary": "Delete a user",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "User ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "User deleted"
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_rod1kutzyy_task-manager-app_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_rod1kutzyy_task-manager-app_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_rod1kutzyy_task-manager-app_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Updates user fields using three-state semantics for each field.\n1. Field is omitted: the value is not changed.\n2. Field has a value: the value is updated.\n3. Field is explicitly null: the value is cleared (set to NULL).\nConstraint: `full_name` cannot be set to null.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Partially update a user",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "User ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "User patch payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_features_users_transport_http.patchUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated user",
                        "schema": {
                            "$ref": "#/definitions/internal_features_users_transport_http.patchUserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_rod1kutzyy_task-manager-app_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_rod1kutzyy_task-manager-app_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_rod1kutzyy_task-manager-app_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_rod1kutzyy_task-manager-app_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_rod1kutzyy_task-manager-app_internal_core_transport_http_response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "error text"
                },
                "message": {
                    "type": "string",
                    "example": "human-readable message"
                }
            }
        },
...
        "internal_features_tasks_transport_http.createTaskRequest": {
            "type": "object",
            "required": [
                "author_user_id",
                "title"
            ],
            "properties": {
                "author_user_id": {
                    "type": "string",
                    "example": "279ee73e-0132-4d5e-9498-cfe6fb43742c"
                },
                "description": {
                    "type": "string",
                    "maxLength": 1000,
                    "minLength": 1,
                    "example": "Morning walk at 06:30"
                },
                "title": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1,
                    "example": "Walk the dog"
                }
            }
        },
        "internal_features_tasks_transport_http.createTaskResponse": {
            "type": "object",
            "properties": {
                "author_user_id": {
                    "type": "string",
                    "example": "279ee73e-0132-4d5e-9498-cfe6fb43742c"
                },
                "completed": {
                    "type": "boolean",
                    "example": true
                },
                "completed_at": {
                    "type": "string",
                    "example": "2026-04-15T09:00:00Z"
                },
                "created_at": {
                    "type": "string",
                    "example": "2026-04-15T08:30:00Z"
                },
                "description": {
                    "type": "string",
                    "example": "Morning walk at 06:30"
                },
                "id": {
                    "type": "string",
                    "example": "b5fc5eb1-65b6-4a28-ac34-612fd14cd39c"
                },
                "title": {
                    "type": "string",
                    "example": "Walk the dog"
                },
                "version": {
                    "type": "integer",
                    "example": 2
                }
            }
        },
        "internal_features_tasks_transport_http.getTaskResponse": {
            "type": "object",
            "properties": {
                "author_user_id": {
                    "type": "string",
                    "example": "279ee73e-0132-4d5e-9498-cfe6fb43742c"
                },
                "completed": {
                    "type": "boolean",
                    "example": true
                },
                "completed_at": {
                    "type": "string",
                    "example": "2026-04-15T09:00:00Z"
                },
                "created_at": {
                    "type": "string",
                    "example": "2026-04-15T08:30:00Z"
                },
                "description": {
                    "type": "string",
                    "example": "Morning walk at 06:30"
                },
                "id": {
                    "type": "string",
                    "example": "b5fc5eb1-65b6-4a28-ac34-612fd14cd39c"
                },
                "title": {
                    "type": "string",
                    "example": "Walk the dog"
                },
                "version": {
                    "type": "integer",
                    "example": 2
                }
            }
        },
        "internal_features_tasks_transport_http.patchTaskRequest": {
            "type": "object",
            "properties": {
                "completed": {
                    "type": "boolean",
                    "example": true
                },
                "description": {
                    "type": "string",
                    "example": "Morning walk at 06:30"
                },
                "title": {
                    "type": "string",
                    "example": "Walk the dog"
                }
            }
        },
        "internal_features_tasks_transport_http.patchTaskResponse": {
            "type": "object",
            "properties": {
                "author_user_id": {
                    "type": "string",
                    "example": "279ee73e-0132-4d5e-9498-cfe6fb43742c"
                },
                "completed": {
                    "type": "boolean",
                    "example": true
                },
                "completed_at": {
                    "type": "string",
                    "example": "2026-04-15T09:00:00Z"
                },
                "created_at": {
                    "type": "string",
                    "example": "2026-04-15T08:30:00Z"
                },
                "description": {
                    "type": "string",
                    "example": "Morning walk at 06:30"
                },
                "id": {
                    "type": "string",
                    "example": "b5fc5eb1-65b6-4a28-ac34-612fd14cd39c"
                },
                "title": {
                    "type": "string",
                    "example": "Walk the dog"
                },
                "version": {
                    "type": "integer",
                    "example": 2
                }
            }
        },
        "internal_features_tasks_transport_http.taskDTOResponse": {
            "type": "object",
            "properties": {
                "author_user_id": {
                    "type": "string",
                    "example": "279ee73e-0132-4d5e-9498-cfe6fb43742c"
                },
                "completed": {
                    "type": "boolean",
                    "example": true
                },
                "completed_at": {
                    "type": "string",
                    "example": "2026-04-15T09:00:00Z"
                },
                "created_at": {
                    "type": "string",
                    "example": "2026-04-15T08:30:00Z"
                },
                "description": {
                    "type": "string",
                    "example": "Morning walk at 06:30"
                },
                "id": {
                    "type": "string",
                    "example": "b5fc5eb1-65b6-4a28-ac34-612fd14cd39c"
                },
                "title": {
                    "type": "string",
                    "example": "Walk the dog"
                },
                "version": {
                    "type": "integer",
                    "example": 2
                }
            }
        },
...
        "internal_features_users_transport_http.getUserResponse": {
            "type": "object",
            "properties": {
                "full_name": {
                    "type": "string",
                    "example": "Ivan Ivanov"
                },
                "id": {
                    "type": "string",
                    "example": "279ee73e-0132-4d5e-9498-cfe6fb43742c"
                },
                "phone_number": {
                    "type": "string",
                    "example": "+79998887766"
                },
                "version": {
                    "type": "integer",
                    "example": 3
                }
            }
        },
        "internal_features_users_transport_http.patchUserRequest": {
            "type": "object",
            "properties": {
                "full_name": {
                    "type": "string",
                    "example": "Ivan Ivanov"
                },
                "phone_number": {
                    "type": "string",
                    "example": "+79998887766"
                }
            }
        },
        "internal_features_users_transport_http.patchUserResponse": {
            "type": "object",
            "properties": {
                "full_name": {
                    "type": "string",
                    "example": "Ivan Ivanov"
                },
                "id": {
                    "type": "string",
                    "example": "279ee73e-0132-4d5e-9498-cfe6fb43742c"
                },
                "phone_number": {
                    "type": "string",
                    "example": "+79998887766"
                },
                "version": {
                    "type": "integer",
                    "example": 3
                }
            }
        },
...
    }
}